update
1. album_get: add year & genre
2. split-album.sh: add failsafe
3. update-*: remove obosolutes

+++ split-album.sh
@@ -33,7 +33,12 @@ for ext in wav flac ape; do
     echo "split ""$file"
 
     # do split 
-    shnsplit -f "$cue" -t '%n.%t' -o flac -O always "$file" || exit $?
+    shnsplit -f "$cue" -t '%n.%t' -o flac -O always "$file" || {
+        # some cue may split failed 
+        cuebreakpoints "$cue" | sed 's/$/0/g' | shnsplit -t '%n' -o flac -O always "$file" 
+    }
+    [ $? -ne 0 ] && exit $?
+
     # remove pregap
     rm -fv 00.pregap.flac || true
     # add tags
@@ -41,10 +46,11 @@ for ext in wav flac ape; do
     # remove cue and its data file
     #rm "$cue" "$file" &&
     # 我们要合并多个CD，track/TRACKTOTAL会导致显示顺序错误 => 按文件名排序
-    #metaflac --preserve-modtime  \
-    #    --remove-tag=TRACKNUMBER \
-    #    --remove-tag=TRACKTOTAL  \
-    #    $(find . -name "*.flac" | grep -v "$file")
+    metaflac --preserve-modtime     \
+        --remove-tag=track          \
+        --remove-tag=TRACKNUMBER    \
+        --remove-tag=TRACKTOTAL     \
+        $(find . -name "*.flac" | grep -v "$file")
     #    #--remove-tag=TITLE       \
     exit
 done
